2/5/2019 3.0.6975.36893 从这个版本开始，老的修改器界面被移除，新的界面更名为QTRHacker 修改器版本号也将采用新的修改器界面的版本号 新界面的开发基本完成，主要功能都已经加入 另外修复了一些问题，QTRHacker.Functions.dll的改动较大，不能保证稳定性，需要测试 说明书已经过时，但仍与新的修改器一起发布，仅供参考

+++ QHackLib/Context.cs
@@ -110,7 +110,7 @@ namespace QHackLib
 
 		public static Context Create(int processID)
 		{
-			string name = Process.GetProcessById((int)processID).ProcessName;
+			string name = Process.GetProcessById(processID).ProcessName;
 			return new Context(name, processID, OpenProcess(PROCESS_ALL_ACCESS, false, processID), name + ".exe");
 		}
